# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Balances and Transactions API",
    "description": "This API provides endpoints to get the current balance and a list of transactions.",
    "version": "1.0.0",
    "servers": [
      {
        "url": "https://uh4goxppjc7stkg24d6fdma4t40wxtly.lambda-url.eu-central-1.on.aws"
      }
    ]
  },
  "paths": {
    "/balances": {
      "get": {
        "security": [
          {
            "xApiKey": []
          }
        ],
        "description": "Get the current balance.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "amount": {
                      "type": "number"
                    },
                    "currency": {
                      "type": "string",
                      "enum": ["EUR"]
                    },
                    "date": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transactions": {
      "get": {
        "security": [
          {
            "xApiKey": []
          }
        ],
        "description": "Get a list of transactions.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transactions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "amount": {
                            "type": "number"
                          },
                          "currency": {
                            "type": "string",
                            "enum": ["EUR"]
                          },
                          "date": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "status": {
                            "type": "string",
                            "enum": ["BOOKED", "PROCESSED", "CANCELLED"]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/historical-balances": {
      "get": {
        "security": [
          {
            "xApiKey": []
          }
        ],
        "description": "Get a list of balances.",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "Start date in DD-MM-YYYY format",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^[0-3][0-9]-[0-1][0-9]-[0-9]{4}$"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End date in DD-MM-YYYY format",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^[0-3][0-9]-[0-1][0-9]-[0-9]{4}$"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order for dates. Can be either 'asc' for ascending or 'desc' for descending",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "amount": {
                        "type": "number"
                      },
                      "currency": {
                        "type": "string",
                        "enum": ["EUR"]
                      },
                      "date": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "xApiKey": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header"
      }
    }
  }
}